# contracts/api/openapi.yaml
openapi: 3.0.3
info:
  title: POC SDD .NET Core API
  description: API para POC de Spec-Driven Development
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1
  - url: https://pocsddnetcoreapi.timeforsoftware.com/api/v1

paths:
  /clients:
    get:
      summary: Obtener todos los clientes
      operationId: getClients
      tags:
        - Clients
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      summary: Crear un nuevo cliente
      operationId: createClient
      tags:
        - Clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'

  /clients/{id}:
    get:
      summary: Obtener cliente por ID
      operationId: getClientById
      tags:
        - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar un cliente
      operationId: updateClient
      tags:
        - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      summary: Eliminar un cliente
      operationId: deleteClient
      tags:
        - Clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente eliminado
        '404':
          $ref: '#/components/responses/NotFound'

  /projects:
    get:
      summary: Obtener todos los proyectos
      operationId: getProjects
      tags:
        - Projects
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filtra los proyectos de un cliente
      responses:
        '200':
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Crear un nuevo proyecto
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Proyecto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /projects/{id}:
    get:
      summary: Obtener proyecto por ID
      operationId: getProjectById
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Proyecto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar un proyecto
      operationId: updateProject
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Proyecto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Eliminar un proyecto
      operationId: deleteProject
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Proyecto eliminado
        '404':
          $ref: '#/components/responses/NotFound'

  /sectors:
    get:
      summary: Obtener el catalogo de sectores (solo lectura)
      operationId: getSectors
      tags:
        - Sectors
      responses:
        '200':
          description: Catalogo de sectores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sector'

  /sectors/{id}:
    get:
      summary: Obtener sector por ID
      operationId: getSectorById
      tags:
        - Sectors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sector encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sector'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        logo:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
          description: URL del logo del cliente (opcional)
        createdAt:
          type: string
          format: date-time

    CreateClientRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 50
        logo:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
          description: URL del logo del cliente (opcional)

    UpdateClientRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 50
        logo:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
          description: URL del logo del cliente (opcional)

    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        sectorId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        technologies:
          type: string
        startDate:
          type: string
          format: date-time
        durationMonths:
          type: integer
          format: int32
          nullable: true
          minimum: 1
          description: Duración en meses (opcional)
        createdAt:
          type: string
          format: date-time

    CreateProjectRequest:
      type: object
      required:
        - clientId
        - sectorId
        - title
        - description
        - technologies
        - startDate
      properties:
        clientId:
          type: string
          format: uuid
        sectorId:
          type: string
          format: uuid
          description: Sector del proyecto (obligatorio, ver /sectors)
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        technologies:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date-time
        durationMonths:
          type: integer
          format: int32
          minimum: 1
          description: Duración en meses (opcional)

    UpdateProjectRequest:
      type: object
      required:
        - clientId
        - sectorId
        - title
        - description
        - technologies
        - startDate
      properties:
        clientId:
          type: string
          format: uuid
        sectorId:
          type: string
          format: uuid
          description: Sector del proyecto (obligatorio, ver /sectors)
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        technologies:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date-time
        durationMonths:
          type: integer
          format: int32
          minimum: 1
          description: Duración en meses (opcional)

    Sector:
      type: object
      description: Catalogo de sectores de solo lectura (7 valores fijos)
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          enum:
            - Administración pública
            - Ingeniería
            - Publicidad
            - Servicios financieros
            - Servicios tecnológicos
            - Transporte
            - Sector inmobiliario

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'